3.0.1: fix setup on macOS
Squashed commit of the following:

commit d0964ea0eca80211b59a100028f5a0e53e9254f6
Author: TitanSnow <sweeto@live.cn>
Date:   Fri Nov 23 22:14:23 2018 +0800

    normalize.py encoding & isfile

commit bf1be642758872a2af6ff3a8a1ad424e7deeb569
Author: TitanSnow <sweeto@live.cn>
Date:   Fri Nov 23 22:10:31 2018 +0800

    upgrade to 3.0.1

commit 0ece79fe2a7fdb7724883672ba30c6b9a13a6157
Author: TitanSnow <sweeto@live.cn>
Date:   Fri Nov 23 21:16:48 2018 +0800

    Move to GitLab CI

    Squashed commit of the following:

    commit bd94bfdba526870b448c7eb9fa00a8c911c9900b
    Author: TitanSnow <sweeto@live.cn>
    Date:   Fri Nov 23 20:45:23 2018 +0800

        -

    commit 87bf4bc929a2b6e484d1595f878690477060803b
    Author: TitanSnow <sweeto@live.cn>
    Date:   Fri Nov 23 20:42:48 2018 +0800

        -

    commit 8859f31abcb21bbf2bff8763b9b4ae69c45bbad3
    Author: TitanSnow <sweeto@live.cn>
    Date:   Fri Nov 23 20:40:18 2018 +0800

        -

    commit 16469be775ed6a5e2c1b6f51fea8a78553852abe
    Author: TitanSnow <sweeto@live.cn>
    Date:   Fri Nov 23 20:38:01 2018 +0800

        -

    commit 33640e34284b7ed63648f34aec7e917f404d1c02
    Author: TitanSnow <sweeto@live.cn>
    Date:   Fri Nov 23 20:35:27 2018 +0800

        -

    commit 56a8f1153ee0ee3afb726efc02fa4fc4a6f8f4b1
    Author: TitanSnow <sweeto@live.cn>
    Date:   Sat Nov 17 19:27:42 2018 +0800

        -

    commit 460f8d093597efd15177a805e7718af04b36ed15
    Author: TitanSnow <sweeto@live.cn>
    Date:   Sat Nov 17 19:25:00 2018 +0800

        -

    commit 9b906a4aa9367572ba11bd566d815ac13cae0975
    Author: TitanSnow <sweeto@live.cn>
    Date:   Sat Nov 17 19:13:30 2018 +0800

        -

    commit d442a5d7d22d4969ed4c00de4f08d9b7a095ed51
    Author: TitanSnow <sweeto@live.cn>
    Date:   Sat Nov 17 19:08:15 2018 +0800

        -

    commit 84dd6c32ad5dcadc0ba95fbb2f5780675f393736
    Author: TitanSnow <sweeto@live.cn>
    Date:   Sat Nov 17 19:03:45 2018 +0800

        -

    commit ef948e43a57b8504916db024a75d9e39db14831f
    Author: TitanSnow <sweeto@live.cn>
    Date:   Sat Nov 17 18:38:30 2018 +0800

        -

    commit 4afdf6570b23c71453d27cf782778e640c9b30b0
    Author: TitanSnow <sweeto@live.cn>
    Date:   Sat Nov 17 18:36:57 2018 +0800

        gitlab ci

commit a14185fbc2726105a1958e971edd3c4355420a59
Author: TitanSnow <sweeto@live.cn>
Date:   Sat Nov 17 18:05:32 2018 +0800

    test on osx

commit 55711845a4c7aec7ff5de879f93ab6d3ab4ce38b
Author: TitanSnow <sweeto@live.cn>
Date:   Sat Nov 17 17:59:18 2018 +0800

    --embedding

commit 0f18b7e7682cc651a8e95fdea9ef46049e139526
Author: TitanSnow <sweeto@live.cn>
Date:   Fri Nov 16 20:35:31 2018 +0800

    MANIFEST.in

commit 7e4bea5f5f4c86e2b8354e920591a084c56a046f
Author: TitanSnow <sweeto@live.cn>
Date:   Fri Nov 16 20:06:59 2018 +0800

    tip when no lua lib

commit b8cfcdc406adbff823b7e78ac065b72b06ff828e
Author: TitanSnow <sweeto@live.cn>
Date:   Fri Nov 16 20:02:45 2018 +0800

    add bundle lua

commit 2382ebeeeb475737543ff9dad696461667d6e7a3
Author: TitanSnow <sweeto@live.cn>
Date:   Fri Nov 16 18:34:49 2018 +0800

    warn for no lua

commit acd6573404b99a71fed179b92f8ecb9f16b508de
Author: TitanSnow <sweeto@live.cn>
Date:   Fri Aug 24 11:49:39 2018 +0800

    fix README.rst encoding

+++ tools/normalize.py
@@ -10,26 +10,27 @@ out = run(('git', 'ls-files'), check=True, universal_newlines=True, stdout=PIPE)
 files = out.splitlines()
 fe = re.compile(r'(?:\r\n|\r)$')
 for fn in files:
-    with open(fn, 'r+') as f:
-        print('processing', fn, '...', end=' ', sep=' ')
-        rst = []
-        modified = False
-        for ln in f:
-            r = fe.sub('\n', ln)
-            if r.rstrip() == '':
-                r = '\n'
-            if r != ln:
+    if os.path.isfile(fn):
+        with open(fn, 'r+', encoding='utf8') as f:
+            print('processing', fn, '...', end=' ', sep=' ')
+            rst = []
+            modified = False
+            for ln in f:
+                r = fe.sub('\n', ln)
+                if r.rstrip() == '':
+                    r = '\n'
+                if r != ln:
+                    modified = True
+                rst.append(r)
+            while rst and rst[-1].rstrip() == '':
+                rst.pop()
                 modified = True
-            rst.append(r)
-        while rst and rst[-1].rstrip() == '':
-            rst.pop()
-            modified = True
-        if rst[-1][-1] != '\n':
-            rst[-1] += '\n'
-        if modified:
-            f.seek(0)
-            f.truncate()
-            f.writelines(rst)
-            print('\x1b[7;1mmodified\x1b[0m')
-        else:
-            print('nothing')
+            if rst[-1][-1] != '\n':
+                rst[-1] += '\n'
+            if modified:
+                f.seek(0)
+                f.truncate()
+                f.writelines(rst)
+                print('\x1b[7;1mmodified\x1b[0m')
+            else:
+                print('nothing')
